+++ R/evaluate_imputed_values.R
@@ -28,7 +28,7 @@
 #' all squared true values}
 #' \item{"NRMSE_tot_sd": RMSE divided by the standard deviation of all true values}
 #' \item{"nr_equal": number of imputed values that are equal to the true values}
-#' \item{"nr_NA": number of values in \code{imp_ds} that are NA (not imputed)}
+#' \item{"nr_NA": number of values in \code{ds_imp} that are NA (not imputed)}
 #' \item{"precision": proportion of imputed values that are equal to the true values}
 #' }
 #' Additionally there are relative versions of bias and MAE implemented. In the
@@ -41,23 +41,23 @@
 #' The argument \code{which_cols} allows the selection of columns
 #' for comparison (see examples).
 #'
-#' If \code{M = NULL} (the default), then all values of \code{imp_ds} and
-#' \code{orig_ds} will be used for the calculation of the evaluation criterion.
+#' If \code{M = NULL} (the default), then all values of \code{ds_imp} and
+#' \code{ds_orig} will be used for the calculation of the evaluation criterion.
 #' If a missing data indicator matrix is given via \code{M}, only the truly
 #' imputed values (values that are marked as missing via \code{M}) will be used
 #' for the calculation. If you want to provide \code{M}, \code{M} must be a
-#' logical matrix of the same dimensions as \code{orig_ds} and missing values
+#' logical matrix of the same dimensions as \code{ds_orig} and missing values
 #' must be coded as TRUE. This is the standard behavior, if you use
 #' \code{\link[base]{is.na}} on a dataset with missing values to generate
 #' \code{M} (see examples). It is possible to combine \code{M} and
 #' \code{which_cols}.
 #'
-#' @param imp_ds a data frame or matrix with imputed values
-#' @param orig_ds a data frame or matrix with original (true) values
+#' @param ds_imp a data frame or matrix with imputed values
+#' @param ds_orig a data frame or matrix with original (true) values
 #' @param which_cols indices or names of columns used for evaluation
 #' @param M NULL (the default) or a missing data indicator matrix; the missing
-#'   data indicator matrix is normally created via \code{is.na(miss_ds)}, where
-#'   \code{miss_ds} is the dataset after deleting values from \code{orig_ds}
+#'   data indicator matrix is normally created via \code{is.na(ds_miss)}, where
+#'   \code{ds_miss} is the dataset after deleting values from \code{ds_orig}
 #'
 #' @export
 #'
@@ -66,35 +66,35 @@
 #'   imputation. \emph{Bioinformatics}, 21(2), 187-198.
 #'
 #' @examples
-#' orig_ds <- data.frame(X = 1:10, Y = 101:110)
-#' miss_ds <- delete_MCAR(orig_ds, 0.3)
-#' imp_ds <- impute_mean(miss_ds)
-#' # compare all values from orig_ds and imp_ds
-#' evaluate_imputed_values(imp_ds, orig_ds)
+#' ds_orig <- data.frame(X = 1:10, Y = 101:110)
+#' ds_miss <- delete_MCAR(ds_orig, 0.3)
+#' ds_imp <- impute_mean(ds_miss)
+#' # compare all values from ds_orig and ds_imp
+#' evaluate_imputed_values(ds_imp, ds_orig)
 #' # compare only the imputed values
-#' M <- is.na(miss_ds)
-#' evaluate_imputed_values(imp_ds, orig_ds, M = M)
+#' M <- is.na(ds_miss)
+#' evaluate_imputed_values(ds_imp, ds_orig, M = M)
 #' # compare only the imputed values in column X
-#' evaluate_imputed_values(imp_ds, orig_ds, M = M, which_cols = "X")
+#' evaluate_imputed_values(ds_imp, ds_orig, M = M, which_cols = "X")
 #'
 #' # NRMSE_tot_mean and NRMSE_col_mean are equal, if columnwise means are equal
-#' orig_ds <- data.frame(X = 1:10, Y = 10:1)
-#' miss_ds <- delete_MCAR(orig_ds, 0.3)
-#' imp_ds <- impute_mean(miss_ds)
-#' evaluate_imputed_values(imp_ds, orig_ds, "NRMSE_tot_mean")
-#' evaluate_imputed_values(imp_ds, orig_ds, "NRMSE_col_mean")
-evaluate_imputed_values <- function(imp_ds, orig_ds, criterion = "RMSE", M = NULL,
-                                    which_cols = seq_len(ncol(imp_ds)),
+#' ds_orig <- data.frame(X = 1:10, Y = 10:1)
+#' ds_miss <- delete_MCAR(ds_orig, 0.3)
+#' ds_imp <- impute_mean(ds_miss)
+#' evaluate_imputed_values(ds_imp, ds_orig, "NRMSE_tot_mean")
+#' evaluate_imputed_values(ds_imp, ds_orig, "NRMSE_col_mean")
+evaluate_imputed_values <- function(ds_imp, ds_orig, criterion = "RMSE", M = NULL,
+                                    which_cols = seq_len(ncol(ds_imp)),
                                     tolerance = sqrt(.Machine$double.eps)) {
-  if (!isTRUE(all.equal(dim(imp_ds), dim(orig_ds)))) {
-    stop("the dimensions of imp_ds and orig_ds must be equal")
+  if (!isTRUE(all.equal(dim(ds_imp), dim(ds_orig)))) {
+    stop("the dimensions of ds_imp and ds_orig must be equal")
   }
 
-  imp_ds <- imp_ds[, which_cols, drop = FALSE]
-  orig_ds <- orig_ds[, which_cols, drop = FALSE]
+  ds_imp <- ds_imp[, which_cols, drop = FALSE]
+  ds_orig <- ds_orig[, which_cols, drop = FALSE]
   if (!is.null(M)) {
     M <- M[, which_cols, drop = FALSE]
   }
 
-  calc_evaluation_criterion(imp_ds, orig_ds, criterion, M, tolerance = tolerance)
+  calc_evaluation_criterion(ds_imp, ds_orig, criterion, M, tolerance = tolerance)
 }