+++ tests/testthat/test-evaluate_imputed_values.R
@@ -22,6 +22,6 @@ test_that("evaluate_imputed_values()", {
 
   expect_error(
     evaluate_imputed_values(df_XY_20_imp, df_XY_100),
-    "the dimensions of imp_ds and orig_ds must be equal"
+    "the dimensions of ds_imp and ds_orig must be equal"
   )
 })